requirements.txt
Django==5.0.6
djangorestframework==3.15.1
Pillow==10.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cleanup==8.1.0 